requirements.txt
Flask==0.10.1
Pillow==3.1.1
PyYAML==3.11
gunicorn==19.4.5
